requirements.txt
requests
pytest
allure-pytest
selenium
